+++ tests/Checkout/ExpressCheckout/SalesChannel/ExpressCreateOrderRouteTest.php
@@ -14,6 +14,7 @@ use Psr\Log\NullLogger;
 use Shopware\Core\Checkout\Cart\SalesChannel\CartService;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
+use Shopware\Core\PlatformRequest;
 use Swag\PayPal\Checkout\ExpressCheckout\SalesChannel\ExpressCreateOrderRoute;
 use Swag\PayPal\Checkout\Payment\Service\VaultTokenService;
 use Swag\PayPal\OrdersApi\Builder\PayPalOrderBuilder;
@@ -21,6 +22,7 @@ use Swag\PayPal\OrdersApi\Builder\Util\AddressProvider;
 use Swag\PayPal\OrdersApi\Builder\Util\AmountProvider;
 use Swag\PayPal\OrdersApi\Builder\Util\ItemListProvider;
 use Swag\PayPal\OrdersApi\Builder\Util\PurchaseUnitProvider;
+use Swag\PayPal\RestApi\V2\Api\Order;
 use Swag\PayPal\RestApi\V2\Resource\OrderResource;
 use Swag\PayPal\Setting\Settings;
 use Swag\PayPal\Test\Helper\CheckoutRouteTrait;
@@ -45,9 +47,12 @@ class ExpressCreateOrderRouteTest extends TestCase
 
     private TestHandler $logger;
 
+    private PayPalClientFactoryMock $clientFactory;
+
     protected function setUp(): void
     {
         $this->logger = new TestHandler();
+        $this->clientFactory = new PayPalClientFactoryMock(new NullLogger());
     }
 
     public function testCreatePayment(): void
@@ -70,11 +75,11 @@ class ExpressCreateOrderRouteTest extends TestCase
         static::assertSame(CreateOrderCapture::ID, $response->getToken());
     }
 
-    public function testCreateWithoutShippingCallback(): void
+    public function testCreateWithLocalEnvironmentActive(): void
     {
         $salesChannelContext = $this->getSalesChannelContext();
 
-        $response = $this->createRoute(true)->createPayPalOrder(new Request(), $salesChannelContext);
+        $response = $this->createRoute([Settings::IS_LOCAL_ENVIRONMENT => true])->createPayPalOrder(new Request(), $salesChannelContext);
 
         static::assertSame(Response::HTTP_OK, $response->getStatusCode());
         static::assertSame(CreateOrderCapture::ID, $response->getToken());
@@ -82,12 +87,81 @@ class ExpressCreateOrderRouteTest extends TestCase
         static::assertTrue($this->logger->hasDebug(['message' => 'Skipped shipping callback due to being disabled in system config']));
     }
 
-    private function createRoute(bool $callbacksDisabled = false): ExpressCreateOrderRoute
+    public function testCreateWithShippingCallbackDisabled(): void
     {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $response = $this->createRoute([Settings::ECS_SHIPPING_CALLBACK_ENABLED => false])->createPayPalOrder(new Request(), $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        static::assertTrue($this->logger->hasDebug('Skipped shipping callback due to being disabled in system config'));
+    }
+
+    public function testCreateShippingCallbackStoreApi(): void
+    {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $router = $this->createMock(RouterInterface::class);
+        $router
+            ->expects(static::once())
+            ->method('generate')
+            ->with('store-api.paypal.express.shipping_callback')
+            ->willReturn('generatedUrl');
+
+        $response = $this->createRoute([], $router)->createPayPalOrder(new Request(), $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $data = $this->clientFactory->getClient()->getData();
+        $order = (new Order())->assign($data);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNotNull($experienceContext->getOrderUpdateCallbackConfig());
+    }
+
+    public function testCreateShippingCallbackStorefront(): void
+    {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $router = $this->createMock(RouterInterface::class);
+        $router
+            ->expects(static::once())
+            ->method('generate')
+            ->with('frontend.paypal.express.shipping_callback')
+            ->willReturn('generatedUrl');
+
+        $request = new Request();
+        $request->attributes->set(PlatformRequest::ATTRIBUTE_ROUTE_SCOPE, ['storefront']);
+
+        $response = $this->createRoute([], $router)->createPayPalOrder($request, $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $data = $this->clientFactory->getClient()->getData();
+        $order = (new Order())->assign($data);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNotNull($experienceContext->getOrderUpdateCallbackConfig());
+    }
+
+    /**
+     * @param array<string, mixed> $systemConfigSettings
+     */
+    private function createRoute(
+        array $systemConfigSettings = [],
+        ?RouterInterface $router = null,
+    ): ExpressCreateOrderRoute {
         $systemConfig = $this->createSystemConfigServiceMock([
             Settings::CLIENT_ID => 'testClientId',
             Settings::CLIENT_SECRET => 'testClientSecret',
-            Settings::IS_LOCAL_ENVIRONMENT => $callbacksDisabled,
+            Settings::ECS_SHIPPING_CALLBACK_ENABLED => true,
+            ...$systemConfigSettings,
         ]);
 
         $priceFormatter = new PriceFormatter();
@@ -108,9 +182,9 @@ class ExpressCreateOrderRouteTest extends TestCase
         return new ExpressCreateOrderRoute(
             $this->getContainer()->get(CartService::class),
             $paypalOrderBuilder,
-            new OrderResource(new PayPalClientFactoryMock(new NullLogger())),
+            new OrderResource($this->clientFactory),
             $systemConfig,
-            $this->createMock(RouterInterface::class),
+            $router ?? $this->createMock(RouterInterface::class),
             new Logger('test', [$this->logger]),
         );
     }